add github releases for mobile
