+++ ci/Jenkinsfile.combined
@@ -22,31 +22,31 @@ pipeline {
         /* load common lib */
         cmn = load('ci/common.groovy')
         /* just for a shorter access */
-        btype = cmn.getBuildType() 
+        btype = cmn.utils.getBuildType() 
       } }
     }
     stage('Build') {
       parallel {
         stage('MacOS') { steps { script {
-          osx = cmn.buildBranch('status-react/combined/desktop-macos')
+          osx    = cmn.ci.Build('status-react/combined/desktop-macos')
         } } }
         stage('Linux') { steps { script {
-          nix = cmn.buildBranch('status-react/combined/desktop-linux')
+          nix    = cmn.ci.Build('status-react/combined/desktop-linux')
         } } }
         stage('Windows') { steps { script {
-          win = cmn.buildBranch('status-react/combined/desktop-windows')
+          win    = cmn.ci.Build('status-react/combined/desktop-windows')
         } } }
         stage('iOS') { steps { script {
-          ios = cmn.buildBranch('status-react/combined/mobile-ios')
+          ios    = cmn.ci.Build('status-react/combined/mobile-ios')
         } } }
         //stage('iOS e2e') { steps { script {
-        //  iose2e = cmn.buildBranch('status-react/combined/mobile-ios-e2e')
+        //  iose2e = cmn.ci.Build('status-react/combined/mobile-ios-e2e')
         //} } }
         stage('Android') { steps { script {
-          apk = cmn.buildBranch('status-react/combined/mobile-android')
+          apk    = cmn.ci.Build('status-react/combined/mobile-android')
         } } }
         stage('Android e2e') { steps { script {
-          apke2e = cmn.buildBranch('status-react/combined/mobile-android-e2e')
+          apke2e = cmn.ci.Build('status-react/combined/mobile-android-e2e')
         } } }
       }
     }
@@ -54,17 +54,17 @@ pipeline {
       steps { script {
         sh('rm -f pkg/*')
         if (btype != 'release') {
-          cmn.copyArts('status-react/combined/desktop-macos', osx.number)
-          cmn.copyArts('status-react/combined/desktop-linux', nix.number)
-          cmn.copyArts('status-react/combined/desktop-windows', win.number)
+          cmn.ci.copyArts(osx)
+          cmn.ci.copyArts(nix)
+          cmn.ci.copyArts(win)
         }
-        cmn.copyArts('status-react/combined/mobile-ios', ios.number)
-        //cmn.copyArts('status-react/combined/mobile-ios-e2e', iose2e.number)
-        cmn.copyArts('status-react/combined/mobile-android', apk.number)
-        cmn.copyArts('status-react/combined/mobile-android-e2e', apke2e.number)
+        cmn.ci.copyArts(ios)
+        //cmn.ci.copyArts(iose2e)
+        cmn.ci.copyArts(apk)
+        cmn.ci.copyArts(apke2e)
         dir('pkg') {
           /* generate sha256 checksums for upload */
-          sh "sha256sum * | tee ${cmn.pkgFilename(btype, 'sha256')}"
+          sh "sha256sum * | tee ${cmn.utils.pkgFilename(btype, 'sha256')}"
           archiveArtifacts('*')
         }
       } }
@@ -76,14 +76,14 @@ pipeline {
           /* mobile */
           Apk: cmn.pkgUrl(apk), Apke2e: cmn.pkgUrl(apke2e),
           iOS: cmn.pkgUrl(ios), /*iOSe2e: cmn.pkgUrl(iose2e),*/
-          Diawi: cmn.getEnv(ios, 'DIAWI_URL'),
+          Diawi: cmn.utils.getEnv(ios, 'DIAWI_URL'),
           /* desktop */
           App: cmn.pkgUrl(nix), Mac: cmn.pkgUrl(osx), Win: cmn.pkgUrl(win),
           /* upload the sha256 checksums file too */
-          SHA: cmn.uploadArtifact(cmn.pkgFind('sha256')),
+          SHA: cmn.utils.uploadArtifact(cmn.utils.pkgFind('sha256')),
         ]
         /* add URLs to the build description */
-        cmn.setBuildDesc(urls)
+        cmn.ci.setBuildDesc(urls)
         /* Create JSON file with newest build URLs */
         switch (btype) {
           /* legacy naming, should have named it nightly.json */
@@ -93,15 +93,17 @@ pipeline {
       } }
     }
     stage('Publish') {
-      when { expression { btype == 'nightly' } }
       steps { script {
-        build('misc/status.im')
+        switch (btype) {
+          case 'nightly': build('misc/status.im'); break
+          case 'release': cmn.gh.publishReleaseMobile(); break
+        }
       } }
     }
     stage('Run e2e') {
       when { expression { btype == 'nightly' } }
       steps { script {
-        e2eApk = cmn.getEnv(apke2e, 'SAUCE_URL')
+        e2eApk = cmn.utils.getEnv(apke2e, 'SAUCE_URL')
         build(
           job: 'end-to-end-tests/status-app-nightly', wait: false,
           parameters: [string(name: 'apk', value: "--apk=${e2eApk}")]
