+++ ci/Jenkinsfile.ios
@@ -38,35 +38,36 @@ pipeline {
           /* Necessary to load methods */
           mobile = load 'ci/mobile.groovy'
           cmn    = load 'ci/common.groovy'
-          print "Running ${cmn.getBuildType()} build!"
-          cmn.abortPreviousRunningBuilds()
+          btype   = cmn.utils.getBuildType()
+          print "Running ${btype} build!"
+          cmn.ci.abortPreviousRunningBuilds()
           /* Read the valid NodeJS version */
-          env.NODE_VERSION = cmn.getToolVersion('node')
+          env.NODE_VERSION = cmn.utils.getToolVersion('node')
           /* Cleanup and Prep */
           nvm(env.NODE_VERSION) {
-            mobile.prep(cmn.getBuildType())
+            mobile.prep(btype)
           }
         }
       }
     }
     stage('Lint') {
       steps {nvm(env.NODE_VERSION) {
-        script { cmn.runLint() }
+        sh 'lein cljfmt check'
       } }
     }
     stage('Tests') {
       steps { nvm(env.NODE_VERSION) {
-        script { cmn.runTests() }
+        sh 'lein test-cljs'
       } }
     }
     stage('Build') {
       steps { nvm(env.NODE_VERSION) {
-        script { mobile.leinBuild('ios') }
+        sh "lein prod-build-ios"
       } }
     }
     stage('Bundle') {
       steps { nvm(env.NODE_VERSION) {
-        script { api = mobile.ios.bundle(cmn.getBuildType()) }
+        script { api = mobile.ios.bundle(btype) }
       } }
     }
     stage('Archive') {
@@ -77,9 +78,9 @@ pipeline {
     stage('Upload') {
       steps {
         script {
-          env.PKG_URL = cmn.uploadArtifact(api)
+          env.PKG_URL = cmn.utils.uploadArtifact(api)
           /* e2e builds get tested in SauceLabs */
-          if (cmn.getBuildType() == 'e2e') {
+          if (btype == 'e2e') {
             env.SAUCE_URL = mobile.ios.uploadToSauceLabs()
           } else {
             env.DIAWI_URL = mobile.ios.uploadToDiawi()
@@ -89,12 +90,12 @@ pipeline {
     }
     stage('Cleanup') {
       steps {
-        script { cmn.clean() }
+        sh 'make clean'
       }
     }
   }
   post {
-    failure { script { load('ci/common.groovy').notifyPRFailure() } }
-    success { script { load('ci/common.groovy').notifyPRSuccess() } }
+    success { script { load('ci/common.groovy').notifyPR(true) } }
+    failure { script { load('ci/common.groovy').notifyPR(false) } }
   }
 }
